refactor: apply PR feedback for radialAxes
