+++ apps/chart/src/component/axis.ts
@@ -1,13 +1,12 @@
 import Component from './component';
 import { ChartState, Options, ViewAxisLabel, AxisData } from '@t/store/store';
 import { makeTickPixelPositions, crispPixel } from '@src/helpers/calculator';
-import { TickModel, LineModel, AxisModels } from '@t/components/axis';
+import { TickModel, LineModel, AxisModels, LabelModel } from '@t/components/axis';
 import { TICK_SIZE } from '@src/brushes/axis';
 import { includes } from '@src/helpers/utils';
 import { getAxisTheme } from '@src/helpers/axes';
 import { AxisTheme } from '@t/theme';
 import { getTitleFontString } from '@src/helpers/style';
-import { LabelModel } from '@t/brush/label';
 
 export enum AxisType {
   X = 'xAxis',
