+++ apps/chart/src/component/axisTitle.ts
@@ -6,7 +6,7 @@ import { includes } from '@src/helpers/utils';
 import { FontTheme } from '@t/theme';
 import { getTitleFontString } from '@src/helpers/style';
 import { getAxisTheme } from '@src/helpers/axes';
-import { LabelModel } from '@t/brush/label';
+import { LabelModel } from '@t/components/axis';
 
 export default class AxisTitle extends Component {
   models!: LabelModel[];
