+++ apps/chart/src/component/axisUsingCenterY.ts
@@ -3,11 +3,10 @@ import Painter from '@src/painter';
 import { AxisType } from '@src/component/axis';
 import { ChartState, Options, CenterYAxisData, ViewAxisLabel } from '@t/store/store';
 import { makeTickPixelPositions, crispPixel } from '@src/helpers/calculator';
-import { TickModel, LineModel, AxisModels } from '@t/components/axis';
+import { TickModel, LineModel, AxisModels, LabelModel } from '@t/components/axis';
 import { AxisTheme } from '@t/theme';
 import { getAxisTheme } from '@src/helpers/axes';
 import { getTitleFontString } from '@src/helpers/style';
-import { LabelModel } from '@t/brush/label';
 
 type CoordinateKey = 'x' | 'y';
