+++ apps/chart/tests/components/radialBarSeries.spec.ts
@@ -120,7 +120,7 @@ describe('radialBar series', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 0,
+        index: 0,
         seriesColor: '#aaaaaa',
         seriesIndex: 0,
       },
@@ -137,7 +137,7 @@ describe('radialBar series', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 0,
+        index: 0,
         seriesColor: '#bbbbbb',
         seriesIndex: 1,
       },
@@ -156,7 +156,7 @@ describe('radialBar series', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 1,
+        index: 1,
         seriesColor: '#aaaaaa',
         seriesIndex: 0,
       },
@@ -173,7 +173,7 @@ describe('radialBar series', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 1,
+        index: 1,
         seriesColor: '#bbbbbb',
         seriesIndex: 1,
       },
@@ -194,7 +194,7 @@ describe('radialBar series', () => {
       clockwise: true,
       drawingStartAngle: -90,
       totalAngle: 360,
-      categoryIndex: 0,
+      index: 0,
       seriesColor: '#aaaaaa',
       seriesIndex: 0,
       data: {
@@ -217,7 +217,7 @@ describe('radialBar series', () => {
       clockwise: true,
       drawingStartAngle: -90,
       totalAngle: 360,
-      categoryIndex: 0,
+      index: 0,
       seriesColor: '#bbbbbb',
       seriesIndex: 1,
       data: {
@@ -240,7 +240,7 @@ describe('radialBar series', () => {
       clockwise: true,
       drawingStartAngle: -90,
       totalAngle: 360,
-      categoryIndex: 1,
+      index: 1,
       seriesColor: '#aaaaaa',
       seriesIndex: 0,
       data: {
@@ -263,7 +263,7 @@ describe('radialBar series', () => {
       clockwise: true,
       drawingStartAngle: -90,
       totalAngle: 360,
-      categoryIndex: 1,
+      index: 1,
       seriesColor: '#bbbbbb',
       seriesIndex: 1,
       data: {
@@ -335,7 +335,7 @@ describe('radar series using null data', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 0,
+        index: 0,
         seriesColor: '#bbbbbb',
         seriesIndex: 1,
       },
@@ -354,7 +354,7 @@ describe('radar series using null data', () => {
         clockwise: true,
         drawingStartAngle: -90,
         totalAngle: 360,
-        categoryIndex: 1,
+        index: 1,
         seriesColor: '#aaaaaa',
         seriesIndex: 0,
       },
