+++ apps/chart/types/components/axis.d.ts
@@ -1,5 +1,22 @@
-import { Point } from '../options';
-import { LabelModel } from '@t/brush/label';
+import { Point, Rect } from '../options';
+import { Nullable, StyleProp, RectStyle } from './series';
+import {
+  StrokeLabelStyleName,
+  StrokeLabelStyle,
+  LabelStyleName,
+  LabelStyle,
+  PathRectStyleName,
+} from '@src/brushes/label';
+
+export type LabelModel = {
+  type: 'label';
+  style?: StyleProp<LabelStyle, LabelStyleName>;
+  stroke?: StyleProp<StrokeLabelStyle, StrokeLabelStyleName>;
+  text: string;
+  opacity?: number;
+  radian?: number;
+  rotationPosition?: Point;
+} & Point;
 
 export type TickModel = {
   type: 'tick';
@@ -23,4 +40,32 @@ export type AxisModels = {
   label: LabelModel[];
   tick: TickModel[];
   axisLine: LineModel[];
+};
+
+export type ArrowDirection = 'top' | 'right' | 'bottom' | 'left';
+type Arrow = {
+  direction: ArrowDirection;
+  points: Point[];
+} & Point;
+
+export type BubbleInfo = Rect & {
+  fill: string;
+  lineWidth?: number;
+  strokeStyle?: string;
+  radius?: number;
+  style?: Nullable<StyleProp<RectStyle, PathRectStyleName>>;
+  radian?: number;
+  direction?: ArrowDirection;
+} & Nullable<Partial<Arrow>>;
+
+export type BubbleLabelModel = {
+  type: 'bubbleLabel';
+  radian?: number;
+  rotationPosition?: Point;
+  bubble: BubbleInfo;
+  label: Point & {
+    text?: string;
+    strokeStyle?: string;
+    style?: StyleProp<LabelStyle, LabelStyleName>;
+  };
 };
