+++ apps/chart/types/components/series.d.ts
@@ -1,13 +1,12 @@
 import { Point, Rect, BezierPoint, BoxSeriesDataType } from '../options';
 import { CircleStyleName, RectStyleName } from '@src/brushes/basic';
 import { TooltipData } from '@t/components/tooltip';
-import { LineModel } from '@t/components/axis';
+import { LineModel, LabelModel } from '@t/components/axis';
 import { SectorStyle, SectorStyleName } from '@src/brushes/sector';
 import { LegendData } from '@t/components/legend';
 import { TreemapSeriesData } from '@t/store/store';
 import Component from '@src/component/component';
 import { BubbleDataLabel } from '@t/theme';
-import { LabelModel } from '@t/brush/label';
 
 export type Nullable<T> = T | null;
 export type StyleProp<T, K> = (T | K)[];
